Integrate ns-3.35 with ns3-mmwave V5.0

+++ src/lte/model/lte-spectrum-value-helper.cc
@@ -98,7 +98,7 @@ static const struct EutraChannelNumbers
 /// number of EUTRA bands
 #define NUM_EUTRA_BANDS (sizeof (g_eutraChannelNumbers) / sizeof (EutraChannelNumbers))
 
-double 
+double
 LteSpectrumValueHelper::GetCarrierFrequency (uint32_t earfcn)
 {
   NS_LOG_FUNCTION (earfcn);
@@ -107,7 +107,7 @@ LteSpectrumValueHelper::GetCarrierFrequency (uint32_t earfcn)
       // FDD downlink
       return GetDownlinkCarrierFrequency (earfcn);
     }
-  else 
+  else
     {
       // either FDD uplink or TDD (for which uplink & downlink have same frequency)
       return GetUplinkCarrierFrequency (earfcn);
@@ -172,12 +172,12 @@ LteSpectrumValueHelper::GetUplinkCarrierFrequency (uint32_t nUl)
   return 1.0e6 * (g_eutraChannelNumbers[i].fUlLow + 0.1 * (nUl - g_eutraChannelNumbers[i].nOffsUl));
 }
 
-double 
-LteSpectrumValueHelper::GetChannelBandwidth (uint16_t transmissionBandwidth)
+double
+LteSpectrumValueHelper::GetChannelBandwidth (uint8_t transmissionBandwidth)
 {
-  NS_LOG_FUNCTION (transmissionBandwidth);
+  NS_LOG_FUNCTION ((uint16_t) transmissionBandwidth);
   switch (transmissionBandwidth)
-    { 
+    {
     case 6:
       return 1.4e6;
     case 15:
@@ -191,7 +191,7 @@ LteSpectrumValueHelper::GetChannelBandwidth (uint16_t transmissionBandwidth)
     case 100:
       return 20.0e6;
     default:
-      NS_FATAL_ERROR ("invalid bandwidth value " << transmissionBandwidth);
+      NS_FATAL_ERROR ("invalid bandwidth value " << (uint16_t) transmissionBandwidth);
     }
 }
 
@@ -209,11 +209,11 @@ struct LteSpectrumModelId
    */
   LteSpectrumModelId (uint32_t f, uint8_t b);
   uint32_t earfcn; ///< EARFCN
-  uint16_t  bandwidth; ///< bandwidth
+  uint8_t  bandwidth; ///< bandwidth
 };
 
 LteSpectrumModelId::LteSpectrumModelId (uint32_t f, uint8_t b)
-  : earfcn (f), 
+  : earfcn (f),
     bandwidth (b)
 {
 }
@@ -230,15 +230,15 @@ operator < (const LteSpectrumModelId& a, const LteSpectrumModelId& b)
 {
   return ( (a.earfcn < b.earfcn) || ( (a.earfcn == b.earfcn) && (a.bandwidth < b.bandwidth) ) );
 }
- 
+
 
 static std::map<LteSpectrumModelId, Ptr<SpectrumModel> > g_lteSpectrumModelMap; ///< LTE spectrum model map
 
 
 Ptr<SpectrumModel>
-LteSpectrumValueHelper::GetSpectrumModel (uint32_t earfcn, uint16_t txBandwidthConfiguration)
+LteSpectrumValueHelper::GetSpectrumModel (uint32_t earfcn, uint8_t txBandwidthConfiguration)
 {
-  NS_LOG_FUNCTION (earfcn << txBandwidthConfiguration);
+  NS_LOG_FUNCTION (earfcn << (uint16_t) txBandwidthConfiguration);
   Ptr<SpectrumModel> ret;
   LteSpectrumModelId key (earfcn, txBandwidthConfiguration);
   std::map<LteSpectrumModelId, Ptr<SpectrumModel> >::iterator it = g_lteSpectrumModelMap.find (key);
@@ -255,7 +255,7 @@ LteSpectrumValueHelper::GetSpectrumModel (uint32_t earfcn, uint16_t txBandwidthC
       Bands rbs;
       for (uint8_t numrb = 0; numrb < txBandwidthConfiguration; ++numrb)
         {
-          BandInfo rb; 
+          BandInfo rb;
           rb.fl = f;
           f += 90e3;
           rb.fc = f;
@@ -270,10 +270,10 @@ LteSpectrumValueHelper::GetSpectrumModel (uint32_t earfcn, uint16_t txBandwidthC
   return ret;
 }
 
-Ptr<SpectrumValue> 
-LteSpectrumValueHelper::CreateTxPowerSpectralDensity (uint32_t earfcn, uint16_t txBandwidthConfiguration, double powerTx, std::vector <int> activeRbs)
+Ptr<SpectrumValue>
+LteSpectrumValueHelper::CreateTxPowerSpectralDensity (uint32_t earfcn, uint8_t txBandwidthConfiguration, double powerTx, std::vector <int> activeRbs)
 {
-  NS_LOG_FUNCTION (earfcn << txBandwidthConfiguration << powerTx << activeRbs);
+  NS_LOG_FUNCTION (earfcn << (uint16_t) txBandwidthConfiguration << powerTx << activeRbs);
 
   Ptr<SpectrumModel> model = GetSpectrumModel (earfcn, txBandwidthConfiguration);
   Ptr<SpectrumValue> txPsd = Create <SpectrumValue> (model);
@@ -295,9 +295,9 @@ LteSpectrumValueHelper::CreateTxPowerSpectralDensity (uint32_t earfcn, uint16_t 
 }
 
 Ptr<SpectrumValue>
-LteSpectrumValueHelper::CreateTxPowerSpectralDensity (uint32_t earfcn, uint16_t txBandwidthConfiguration, double powerTx, std::map<int, double> powerTxMap, std::vector <int> activeRbs)
+LteSpectrumValueHelper::CreateTxPowerSpectralDensity (uint32_t earfcn, uint8_t txBandwidthConfiguration, double powerTx, std::map<int, double> powerTxMap, std::vector <int> activeRbs)
 {
-  NS_LOG_FUNCTION (earfcn << txBandwidthConfiguration << activeRbs);
+  NS_LOG_FUNCTION (earfcn << (uint16_t) txBandwidthConfiguration << activeRbs);
 
   Ptr<SpectrumModel> model = GetSpectrumModel (earfcn, txBandwidthConfiguration);
   Ptr<SpectrumValue> txPsd = Create <SpectrumValue> (model);
@@ -333,9 +333,9 @@ LteSpectrumValueHelper::CreateTxPowerSpectralDensity (uint32_t earfcn, uint16_t 
 }
 
 Ptr<SpectrumValue>
-LteSpectrumValueHelper::CreateUlTxPowerSpectralDensity (uint16_t earfcn, uint16_t txBandwidthConfiguration, double powerTx, std::vector <int> activeRbs)
+LteSpectrumValueHelper::CreateUlTxPowerSpectralDensity (uint16_t earfcn, uint8_t txBandwidthConfiguration, double powerTx, std::vector <int> activeRbs)
 {
-  NS_LOG_FUNCTION (earfcn << txBandwidthConfiguration << powerTx << activeRbs);
+  NS_LOG_FUNCTION (earfcn << (uint16_t) txBandwidthConfiguration << powerTx << activeRbs);
 
   Ptr<SpectrumModel> model = GetSpectrumModel (earfcn, txBandwidthConfiguration);
   Ptr<SpectrumValue> txPsd = Create <SpectrumValue> (model);
@@ -357,9 +357,9 @@ LteSpectrumValueHelper::CreateUlTxPowerSpectralDensity (uint16_t earfcn, uint16_
 }
 
 Ptr<SpectrumValue>
-LteSpectrumValueHelper::CreateNoisePowerSpectralDensity (uint32_t earfcn, uint16_t txBandwidthConfiguration, double noiseFigure)
+LteSpectrumValueHelper::CreateNoisePowerSpectralDensity (uint32_t earfcn, uint8_t txBandwidthConfiguration, double noiseFigure)
 {
-  NS_LOG_FUNCTION (earfcn << txBandwidthConfiguration << noiseFigure);
+  NS_LOG_FUNCTION (earfcn << (uint16_t) txBandwidthConfiguration << noiseFigure);
   Ptr<SpectrumModel> model = GetSpectrumModel (earfcn, txBandwidthConfiguration);
   return CreateNoisePowerSpectralDensity (noiseFigure, model);
 }
