+++ src/lte/test/lte-test-downlink-power-control.h
@@ -57,7 +57,7 @@ public:
  * \ingroup lte-test
  * \ingroup tests
  *
- * \brief Test SINR calculation in the downlink when power control is used. 
+ * \brief Test SINR calculation in the downlink when power control is used.
  * Test if the difference in power levels are corresponding to the estamated values.
  */
 class LteDownlinkPowerControlSpectrumValueTestCase : public TestCase
@@ -69,13 +69,13 @@ public:
  * \param name the reference name
  * \param earfcn the EARFCN
  * \param bw the bandwidth
- * \param powerTx 
+ * \param powerTx
  * \param powerTxMap
  * \param activeRbs
  * \param expected the expected Tx Power Spectral Density
  */
   LteDownlinkPowerControlSpectrumValueTestCase (std::string name,
-                                                uint16_t earfcn, uint16_t bw, double powerTx,
+                                                uint16_t earfcn, uint8_t bw, double powerTx,
                                                 std::map<int, double> powerTxMap, std::vector <int> activeRbs,
                                                 SpectrumValue& expected);
   virtual ~LteDownlinkPowerControlSpectrumValueTestCase ();
@@ -92,7 +92,7 @@ private:
  * \ingroup tests
  *
  * \brief Test SINR calculation in the downlink when the power control is used.
- * Test the power control by comparing the downlink data and ctrl power 
+ * Test the power control by comparing the downlink data and ctrl power
  * difference with the estimated value based on the specified change in power.
  */
 class LteDownlinkPowerControlTestCase : public TestCase
@@ -121,9 +121,9 @@ private:
  * \ingroup lte-test
  * \ingroup tests
  *
- * \brief Test if RRC connection reconfiguration messages are properly 
+ * \brief Test if RRC connection reconfiguration messages are properly
  * generated upon the change in the downlink power.
- * 
+ *
  */
 class LteDownlinkPowerControlRrcConnectionReconfigurationTestCase : public TestCase
 {
